cleaned up old schemas and their associated logic
