+++ code/server/app.js
@@ -47,7 +47,6 @@ app.use('/locations', routes.locations)
 app.use('/timings', routes.timings)
 app.use('/khateebs', routes.khateebs)
 app.use('/announcements', routes.announcements)
-app.use('/settings', routes.settings)
 app.use('/institutions', routes.institutions)
 app.use('/institutionAdmins', routes.institutionAdmins)
 app.use('/misc', routes.misc)
