+++ code/server/database/main.js
@@ -6,7 +6,6 @@ const institution = require('./schemas/institution.js')
 const timing = require('./schemas/timing.js')
 const user = require('./schemas/user.js')
 const announcement = require('./schemas/announcement.js')
-const setting = require('./schemas/setting.js')
 const notification = require('./schemas/notification.js')
 const shortenedURL = require('./schemas/shortenedURL.js')
 const verificationCode = require('./schemas/verificationCode.js')
@@ -18,7 +17,6 @@ const models = {
     locations: mongoose.model('location', location),
     users: mongoose.model('user', user),
     announcements: mongoose.model('announcement', announcement),
-    settings: mongoose.model('setting', setting),
     notifications: mongoose.model('notification', notification),
     shortenedURLs: mongoose.model('shortenedURL', shortenedURL),
     verificationCodes: mongoose.model('verificationCode', verificationCode),
